refactor: move Provider function into external file
